Revert "Merge branch 'pad-before-core-scattering' into backend-agnostic-reshape"
This reverts commit c31ef30d6250ce9107224bdbf69cdc12a6623ac1, reversing
changes made to a7a8f75e43bae49cc4cf3785b447bb1bb06c4a19.
