+++ kymatio/frontend/numpy_frontend.py
@@ -4,6 +4,10 @@ class ScatteringNumPy:
     def __init__(self):
         self.frontend_name = 'numpy'
 
+    def scattering(self, x):
+        """ This function should compute the scattering transform."""
+        raise NotImplementedError
+
     def __call__(self, x):
         """This method is an alias for `scattering`."""
