+++ kymatio/scattering1d/frontend/numpy_frontend.py
@@ -1,6 +1,7 @@
 import warnings
 
 from ...frontend.numpy_frontend import ScatteringNumPy
+from ..core.scattering1d import scattering1d
 from .base_frontend import ScatteringBase1D
 
 
@@ -13,6 +14,30 @@ class ScatteringNumPy1D(ScatteringNumPy, ScatteringBase1D):
         ScatteringBase1D._instantiate_backend(self, 'kymatio.scattering1d.backend.')
         ScatteringBase1D.build(self)
         ScatteringBase1D.create_filters(self)
+
+    def scattering(self, x):
+        ScatteringBase1D._check_runtime_args(self)
+        ScatteringBase1D._check_input(self, x)
+        x_shape = self.backend.shape(x)
+        batch_shape, signal_shape = x_shape[:-1], x_shape[-1:]
+        x = self.backend.reshape_input(x, signal_shape)
+
+        S = scattering1d(x, self.backend, self.psi1_f, self.psi2_f,
+                         self.phi_f, max_order=self.max_order, average=self.average, pad_left=self.pad_left,
+                         pad_right=self.pad_right, ind_start=self.ind_start, ind_end=self.ind_end,
+                         oversampling=self.oversampling)
+
+        for n, path in enumerate(S):
+            S[n]['coef'] = self.backend.reshape_output(
+                path['coef'], batch_shape, n_kept_dims=1)
+
+        if self.out_type == 'array':
+            return self.backend.concatenate([path['coef'] for path in S], dim=-2)
+        elif self.out_type == 'dict':
+            return {path['n']: path['coef'] for path in S}
+        elif self.out_type == 'list':
+            return S
+
 
 ScatteringNumPy1D._document()
